requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.1
fastapi==0.103.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.0.308
langsmith==0.0.41
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
onnxruntime==1.16.0
openai==0.28.1
overrides==7.4.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyPDF2==3.0.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
